+++ grpc/middleware/correlationid/rand_test.go
@@ -1,0 +1,16 @@
+package correlationid_test
+
+import (
+	"testing"
+
+	"github.com/goph/fxt/grpc/middleware/correlationid"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestRandGenerator_Generate(t *testing.T) {
+	generator := correlationid.NewRandGenerator(20)
+
+	id := generator.Generate()
+
+	assert.Len(t, id, 20)
+}